TBA Classwork-3
This code goes through initial steps such as importing necessary libraries like Pandas for data handling and NLTK for natural language processing tasks. Subsequent sections introduce regular expressions for text cleaning, and the setup of a text processing pipeline that includes functions like tokenization and stop word removal. Practical application of these techniques is demonstrated on a dataset named 'un-general-debates-blueprint.csv'. Each cell in the notebook is well-documented with markdown that explains the purpose and the approach of the code.