TBA Assignment-3
The code contains reading a CSV file 'emotion_sentimen_dataset.csv', into a Pandas Data frame. The notebook provides  how to manipulate and prepare textual data for analysis, which includes cleaning and preprocessing steps. The notebook also demonstrates the creation and application of a text processing pipeline. This pipeline includes several functions aimed at transforming raw text into a format suitable for analysis. These preprocessing steps to the dataset is shown through the execution of the pipeline, with intermediate outputs displayed to demonstrate the effects of each preprocessing step.