TBA_Mid_term_exam
This code file starts with reading the dataset named tweet_emotions.csv then I performed  all the statistical operations to perform cleaning and analysis of data and form various kinds of plots to understand the data better. In continuation with this I also did some tasks like creating a pipeline, remove stop words and word frequency analysis and word clouds. strings which we will encounter in our day-to-day life  and also write regex to identify them and sort is also performed. with the dataset I also performed sentiment analysis of data.