TBA Assignment-5
The notebook begins by importing necessary libraries for data preprocessing and natural language processing, including the setup of the NLTK opinion lexicon, which plays a central role in the sentiment analysis. It then moves on to reading and preparing the data from the JSON file, followed by the application of the Bing Liu lexicon to calculate sentiment scores for the dataset. Each text entry's sentiment is evaluated based on the lexicon's categorization of words into positive and negative sentiments. The notebook also includes visualizations that display the distribution of sentiments, offering insights into the overall sentiment trends within the data. Each step of the process is thoroughly documented with markdown cells that explain the purpose and implementation of the code, ensuring the notebook is both informative and accessible for users seeking to understand or replicate the sentiment analysis.